added pair management

+++ Flashcards/files/fileManager.cs
@@ -11,20 +11,20 @@ namespace Flashcards.files
     public class fileManager
     {
         protected string fileName;
-        protected string[] fileContent;                                         
+        protected string[] fileContent;                                         //variable for saving into file
         List<string> fileContents = new List<string>();                         //list of pairs
-        public string folderLocation = "profiles/";
+        public string folderLocation = "profiles/";                             //main folder
 
-        public void saveFile(string fileName,string profile)
+        public void saveTextFile(string fileName,string profile)
         {
             if (profile!=null &&(profile.Length > 0 && fileName.Length>0))
             {
-                if (!doesFileExist(fileName, profile))
+                if (!doesFileExist(fileName+".txt", profile))
                 {
                     string filePath = this.folderLocation + profile + "/" + fileName + ".txt";
                     FileInfo file = new FileInfo(filePath);
                     file.Directory.Create();
-                    File.WriteAllLines(@filePath, this.fileContents.ToArray());
+                    file.Create();
                 }
                 else
                 {
@@ -35,17 +35,77 @@ namespace Flashcards.files
             {
                 MessageBox.Show("Invalid profile name or file name.");
             }
-            //System.IO.File.WriteAllLines(@"C:\Users\Public\TestFolder\WriteLines.txt", lines);
+        }
+
+        public void saveWordPairsFromList(string fileName, string profile, listManager LM)
+        {
+            if (profile != null && (profile.Length > 0 && fileName.Length > 0))
+            {
+                if (doesFileExist(fileName, profile))
+                {
+                    string filePath = this.folderLocation + profile + "/" + fileName;
+                    foreach (pairWords PW in LM.getList())
+                    {
+                        this.addToContent(PW.getPairTogetherString());
+                    }
+                    File.WriteAllLines(@filePath, this.fileContents.ToArray());
+                }
+                else
+                {
+                    MessageBox.Show("File doesn't exists. Try to recreate file list.");
+                }
+            }
+            else
+            {
+                MessageBox.Show("Invalid profile name or file name.");
+            }
+        }
+
+        public List<pairWords> getWordPairsFromFile(string fileName, string profile)
+        {
+            if (profile != null && (profile.Length > 0 && fileName.Length > 0))
+            {
+                if (doesFileExist(fileName, profile))
+                {
+                    string filePath = this.folderLocation + profile + "/" + fileName;
+                    List<pairWords> returnedList= new List<pairWords>();
+                    string[] wordsTable;
+                    FileInfo file = new FileInfo(filePath);
+                    this.fileContents.Clear();
+                    try
+                    {
+                        var lines = File.ReadAllLines(filePath);
+                        foreach (var line in lines)
+                        {
+                            wordsTable=line.Split('@');
+                            if (wordsTable.Length == 2)
+                            {
+                                returnedList.Add(new pairWords(wordsTable[0],wordsTable[1]));
+                            }
+                        }
+                    }
+                    catch(Exception e)
+                    {
+                        MessageBox.Show("File is used right now. "+ e.ToString());
+                    }
+                    return returnedList;
+                }
+                else
+                {
+                    MessageBox.Show("File doesn't exists. Check if file is not corrupted.");
+                    return null;
+                }
+            }
+            else
+            {
+                MessageBox.Show("Invalid profile name or file name.");
+                return null;
+            }
         }
 
         public void addToContent(string stringToAdd)
         {
            fileContents.Add(stringToAdd);
-        }
-
-        public string getContents()
-        {
-            return string.Join("\n",fileContents.ToArray());
         }
 
         public void createNewProfile(string profileName)
@@ -79,10 +139,11 @@ namespace Flashcards.files
         {
             return profiles.Contains(this.folderLocation + profileName);
         }
+
         public bool doesFileExist(string fileName, string profile)
         {
             List<string> fileList = getFiles(profile);
-            return fileList.Contains(fileName+".txt");
+            return fileList.Contains(fileName);
         }
 
         public List<string> readFromFile(string fileName, string profile)
@@ -90,10 +151,17 @@ namespace Flashcards.files
             string filePath = this.folderLocation + profile+"/"+fileName;
             FileInfo file = new FileInfo(filePath);
             this.fileContents.Clear();
-            var lines = File.ReadAllLines(filePath);
-            foreach(var line in lines)
+            try
             {
-                this.fileContents.Add(line);
+                var lines = File.ReadAllLines(filePath);
+                foreach (var line in lines)
+                {
+                    this.fileContents.Add(line);
+                }
+            }
+            catch (Exception e)
+            {
+                MessageBox.Show("File is used right now. "+e.ToString());
             }
             return this.fileContents;
         }
@@ -104,8 +172,11 @@ namespace Flashcards.files
             {
                 if(File.Exists(this.folderLocation + profile + "/" + fileName))
                 {
-                    MessageBox.Show("Deleting file " + fileName + " from profile " + profile);
-                    File.Delete(this.folderLocation + profile + "/" + fileName);
+                    MessageBoxResult result= MessageBox.Show("Confirm: Deleting file " + fileName + " from profile " + profile+"?","Delete File",MessageBoxButton.YesNo);
+                    if (result == MessageBoxResult.Yes)
+                    {
+                        File.Delete(this.folderLocation + profile + "/" + fileName);
+                    }
                 } 
             }
             else
